+++ tests/testthat/test_auk-filter.r
@@ -2,14 +2,13 @@ context("auk_filter")
 
 test_that("auk_filter filter an ebd", {
   skip_on_cran()
-  skip_on_os("windows")
 
   # set up filters
   f <- system.file("extdata/ebd-sample.txt", package = "auk")
   filters <- auk_ebd(f) %>%
-    auk_species(species = c("Gray Jay", "Blue Jay")) %>%
+    auk_species(species = c("Canada Jay", "Blue Jay")) %>%
     auk_country(country = c("US", "Canada")) %>%
-    auk_extent(extent = c(-100, 37, -80, 52)) %>%
+    auk_bbox(bbox = c(-100, 37, -80, 52)) %>%
     auk_date(date = c("2012-01-01", "2012-12-31")) %>%
     auk_time(start_time = c("06:00", "09:00")) %>%
     auk_duration(duration = c(0, 120)) %>%
@@ -34,10 +33,10 @@ test_that("auk_filter filter an ebd", {
   expect_true(all(ebd$all_species_reported))
   expect_true(all(ebd$time_observations_started >= filters$filters$time[1]))
   expect_true(all(ebd$time_observations_started <= filters$filters$time[2]))
-  expect_true(all(ebd$longitude >= filters$filters$extent[1]))
-  expect_true(all(ebd$longitude <= filters$filters$extent[3]))
-  expect_true(all(ebd$latitude >= filters$filters$extent[2]))
-  expect_true(all(ebd$latitude <= filters$filters$extent[4]))
+  expect_true(all(ebd$longitude >= filters$filters$bbox[1]))
+  expect_true(all(ebd$longitude <= filters$filters$bbox[3]))
+  expect_true(all(ebd$latitude >= filters$filters$bbox[2]))
+  expect_true(all(ebd$latitude <= filters$filters$bbox[4]))
   
   # filter again
   tmp <- tempfile()
@@ -64,7 +63,6 @@ test_that("auk_filter filter an ebd", {
 
 test_that("auk_filter filter sampling and ebd files", {
   skip_on_cran()
-  skip_on_os("windows")
 
   # set up filters
   f <- system.file("extdata/zerofill-ex_ebd.txt", package = "auk")
@@ -99,7 +97,6 @@ test_that("auk_filter filter sampling and ebd files", {
 
 test_that("auk_filter turn off filtering of sampling event data", {
   skip_on_cran()
-  skip_on_os("windows")
 
   # set up filters
   f <- system.file("extdata/zerofill-ex_ebd.txt", package = "auk")
@@ -117,12 +114,11 @@ test_that("auk_filter turn off filtering of sampling event data", {
 
 test_that("auk_filter won't overwrite files", {
   skip_on_cran()
-  skip_on_os("windows")
 
   # set up filters
   f <- system.file("extdata/ebd-sample.txt", package = "auk")
   filters <- auk_ebd(f) %>%
-    auk_species(species = c("Gray Jay", "Blue Jay"))
+    auk_species(species = c("Canada Jay", "Blue Jay"))
 
   # run first time
   tmp <- tempfile()
@@ -138,7 +134,7 @@ test_that("auk_filter can save awk file on any system", {
   # set up filters
   f <- system.file("extdata/ebd-sample.txt", package = "auk")
   filters <- auk_ebd(f) %>%
-    auk_species(species = c("Gray Jay", "Blue Jay"))
+    auk_species(species = c("Canada Jay", "Blue Jay"))
 
   # run first time
   tmp <- tempfile()
@@ -153,7 +149,6 @@ test_that("auk_filter can save awk file on any system", {
 
 test_that("auk_filter filter an auk_sampling object", {
   skip_on_cran()
-  skip_on_os("windows")
   
   # set up filters
   f <- system.file("extdata/zerofill-ex_sampling.txt", package = "auk")
@@ -183,7 +178,6 @@ test_that("auk_filter filter an auk_sampling object", {
 
 test_that("auk_filter works with wildcard dates", {
   skip_on_cran()
-  skip_on_os("windows")
   
   # set up filters
   f <- system.file("extdata/ebd-sample.txt", package = "auk")